feat: 【APM 策略模板】unapply 接口 #1010158081128177462 # Reviewed, transaction id: 62310

+++ bkmonitor/packages/apm_web/strategy/views.py
@@ -214,13 +214,11 @@ class StrategyTemplateViewSet(GenericViewSet):
             service_name__in=entity_set.service_names,
             strategy_template_id__in=self.query_data["strategy_template_ids"],
         )
+        ids: list[int] = list(strategy_instance_qs.values_list("strategy_id", flat=True))
+        if not ids:
+            return Response({})
         # 先删除策略
-        resource.strategies.delete_strategy_v2(
-            {
-                "bk_biz_id": self.query_data["bk_biz_id"],
-                "ids": list(strategy_instance_qs.values_list("strategy_id", flat=True)),
-            }
-        )
+        resource.strategies.delete_strategy_v2({"bk_biz_id": self.query_data["bk_biz_id"], "ids": ids})
         # 再删除策略实例
         strategy_instance_qs.delete()
